+++ miranda/plugins/db3x/dbsettings.c
@@ -103,7 +103,7 @@ static char* InsertCachedSetting( const char* szName, size_t cbNameLen, int inde
 static char* GetCachedSetting(const char *szModuleName,const char *szSettingName,int settingNameLen)
 {
 	static char *lastsetting = NULL;
-	int moduleNameLen = strlen(szModuleName),index;
+	int moduleNameLen = (int)strlen(szModuleName),index;
 	char *szFullName = (char*)alloca(moduleNameLen+settingNameLen+3);
 
 	strcpy(szFullName+1,szModuleName);
@@ -235,7 +235,7 @@ static __inline int GetContactSettingWorker(HANDLE hContact,DBCONTACTGETSETTING 
 
 	if ((!dbcgs->szSetting) || (!dbcgs->szModule))
 		return 1;
-	settingNameLen=strlen(dbcgs->szSetting);
+	settingNameLen=(int)strlen(dbcgs->szSetting);
 
 	EnterCriticalSection(&csDbAccess);
 
@@ -252,7 +252,7 @@ static __inline int GetContactSettingWorker(HANDLE hContact,DBCONTACTGETSETTING 
 				if ( isStatic ) {
 					int cbLen = 0;
 					if ( pCachedValue->pszVal != NULL )
-						cbLen = strlen( pCachedValue->pszVal );
+						cbLen = (int)strlen( pCachedValue->pszVal );
 
 					cbOrigLen--;
 					dbcgs->pValue->pszVal = cbOrigPtr;
@@ -552,8 +552,8 @@ static INT_PTR WriteContactSetting(WPARAM wParam,LPARAM lParam)
 	if ((!dbcws->szModule) || (!dbcws->szSetting) || ((dbcws->value.type == DBVT_ASCIIZ || dbcws->value.type == DBVT_UTF8 )&& dbcws->value.pszVal == NULL) || (dbcws->value.type == DBVT_BLOB && dbcws->value.pbVal == NULL) )
 		return 1;
 	// the db format can't tolerate more than 255 bytes of space (incl. null) for settings+module name
-	settingNameLen=strlen(dbcws->szSetting);
-	moduleNameLen=strlen(dbcws->szModule);
+	settingNameLen=(int)strlen(dbcws->szSetting);
+	moduleNameLen=(int)strlen(dbcws->szModule);
 	if ( settingNameLen > 0xFE )
 	{
 		#ifdef _DEBUG
@@ -572,7 +572,7 @@ static INT_PTR WriteContactSetting(WPARAM wParam,LPARAM lParam)
 	// the db can not tolerate strings/blobs longer than 0xFFFF since the format writes 2 lengths
 	switch( dbcws->value.type ) {
 	case DBVT_ASCIIZ:		case DBVT_BLOB:	case DBVT_UTF8:
-		{	int len = ( dbcws->value.type != DBVT_BLOB ) ? strlen(dbcws->value.pszVal) : dbcws->value.cpbVal;
+		{	int len = ( dbcws->value.type != DBVT_BLOB ) ? (int)strlen(dbcws->value.pszVal) : dbcws->value.cpbVal;
 			if ( len >= 0xFFFF ) {
 				#ifdef _DEBUG
 					OutputDebugString("WriteContactSetting() writing huge string/blob, rejecting ( >= 0xFFFF ) \n");
@@ -627,7 +627,7 @@ static INT_PTR WriteContactSetting(WPARAM wParam,LPARAM lParam)
 	ofsSettingsGroup=GetSettingsGroupOfsByModuleNameOfs(&dbc,ofsContact,ofsModuleName);
 	if(ofsSettingsGroup==0) {  //module group didn't exist - make it
 		if(dbcws->value.type&DBVTF_VARIABLELENGTH) {
-		  if(dbcws->value.type==DBVT_ASCIIZ || dbcws->value.type==DBVT_UTF8) bytesRequired=strlen(dbcws->value.pszVal)+2;
+		  if(dbcws->value.type==DBVT_ASCIIZ || dbcws->value.type==DBVT_UTF8) bytesRequired=(int)strlen(dbcws->value.pszVal)+2;
 		  else if(dbcws->value.type==DBVT_BLOB) bytesRequired=dbcws->value.cpbVal+2;
 		}
 		else bytesRequired=dbcws->value.type;
@@ -692,7 +692,7 @@ static INT_PTR WriteContactSetting(WPARAM wParam,LPARAM lParam)
 					case DBVT_WORD: DBWrite(ofsBlobPtr,&dbcws->value.wVal,2); break;
 					case DBVT_DWORD: DBWrite(ofsBlobPtr,&dbcws->value.dVal,4); break;
 					case DBVT_UTF8:
-					case DBVT_ASCIIZ: DBWrite(ofsBlobPtr+2,dbcws->value.pszVal,strlen(dbcws->value.pszVal)); break;
+					case DBVT_ASCIIZ: DBWrite(ofsBlobPtr+2,dbcws->value.pszVal,(int)strlen(dbcws->value.pszVal)); break;
 					case DBVT_BLOB: DBWrite(ofsBlobPtr+2,dbcws->value.pbVal,dbcws->value.cpbVal); break;
 				}
 				//quit
@@ -708,7 +708,7 @@ static INT_PTR WriteContactSetting(WPARAM wParam,LPARAM lParam)
 	//pBlob already points to end of list
 	//see if it fits
 	if(dbcws->value.type&DBVTF_VARIABLELENGTH) {
-	  if(dbcws->value.type==DBVT_ASCIIZ || dbcws->value.type==DBVT_UTF8) bytesRequired=strlen(dbcws->value.pszVal)+2;
+	  if(dbcws->value.type==DBVT_ASCIIZ || dbcws->value.type==DBVT_UTF8) bytesRequired=(int)strlen(dbcws->value.pszVal)+2;
 	  else if(dbcws->value.type==DBVT_BLOB) bytesRequired=dbcws->value.cpbVal+2;
 	}
 	else bytesRequired=dbcws->value.type;
@@ -765,7 +765,7 @@ static INT_PTR WriteContactSetting(WPARAM wParam,LPARAM lParam)
 		case DBVT_DWORD: DBWrite(ofsBlobPtr,&dbcws->value.dVal,4); MoveAlong(4); break;
 		case DBVT_UTF8:
 		case DBVT_ASCIIZ:
-			{	int len=strlen(dbcws->value.pszVal);
+			{	int len=(int)strlen(dbcws->value.pszVal);
 				DBWrite(ofsBlobPtr,&len,2);
 				DBWrite(ofsBlobPtr+2,dbcws->value.pszVal,len);
 				MoveAlong(2+len);
@@ -795,7 +795,7 @@ static INT_PTR DeleteContactSetting(WPARAM wParam,LPARAM lParam)
 	DWORD ofsModuleName,ofsSettingsGroup,ofsBlobPtr;
 	struct DBContactSettings dbcs;
 	PBYTE pBlob;
-	int settingNameLen=strlen(dbcgs->szSetting),bytesRemaining;
+	int settingNameLen=(int)strlen(dbcgs->szSetting),bytesRemaining;
 	char* szCachedSettingName;
 	WPARAM saveWparam = wParam;
